read_serial_data/requirements.txt
numpy==1.23.1
pyserial==3.5
